Add an utility module to build and render problems
Test coverage is still insufficient but I need to make my way through
a beta for use in Gradle.

This commit adds a new module which responsibility is to offer simple
builders for problems and solutions. It is in general encouraged to
write your own builders for strongly typed problems, instead of using
the generic types (this makes the code easier to read).

However, for simple cases, you might want to be able to generate
simple error messages, and possibly throw them as exceptions: this is
what this module does.